Единая инициализация Telegram-сессии
