+++ pkg/telegram/channel_duplicate/channel_duplicate.go
@@ -10,7 +10,6 @@ import (
 
 	"atg_go/pkg/storage"
 	base "atg_go/pkg/telegram/module"
-	accountmutex "atg_go/pkg/telegram/module/account_mutex"
 
 	"github.com/gotd/td/tg"
 )
@@ -22,42 +21,20 @@ type channelInfo struct {
 	target *tg.Channel
 }
 
-// Start запускает отслеживание донорских каналов и пересылку постов.
-func Start(db *storage.DB) {
-	go func() {
-		if err := run(db); err != nil {
-			log.Printf("[CHANNEL DUPLICATE] остановлено: %v", err)
-		}
-	}()
-}
-
-// run выполняет основную работу по пересылке постов.
-func run(db *storage.DB) error {
-	rand.Seed(time.Now().UnixNano())
-	accounts, err := db.GetMonitoringAccounts()
-	if err != nil {
-		return err
-	}
-	if len(accounts) == 0 {
-		return fmt.Errorf("нет аккаунтов для мониторинга")
-	}
-	acc := accounts[0]
-	if err := accountmutex.LockAccount(acc.ID); err != nil {
-		return err
-	}
-	defer accountmutex.UnlockAccount(acc.ID)
-
+// Connect присоединяет модуль дублирования каналов к существующему клиенту Telegram.
+// Модуль использует уже готовые api и диспетчер, не открывая сессию повторно.
+func Connect(ctx context.Context, api *tg.Client, dispatcher *tg.UpdateDispatcher, db *storage.DB, accountID int) {
 	dups, err := db.GetChannelDuplicates()
 	if err != nil {
-		return err
+		log.Printf("[CHANNEL DUPLICATE] получение списка дубликатов: %v", err)
+		return
 	}
 	if len(dups) == 0 {
-		return fmt.Errorf("нет каналов для дублирования")
+		log.Printf("[CHANNEL DUPLICATE] нет каналов для дублирования")
+		return
 	}
 
-	dispatcher := tg.NewUpdateDispatcher()
 	chMap := make(map[int64]channelInfo)
-	var api *tg.Client
 
 	dispatcher.OnNewChannelMessage(func(ctx context.Context, e tg.Entities, upd *tg.UpdateNewChannelMessage) error {
 		msg, ok := upd.Message.(*tg.Message)
@@ -101,71 +78,60 @@ func run(db *storage.DB) error {
 		return nil
 	})
 
-	client, err := base.Modf_AccountInitialization(acc.ApiID, acc.ApiHash, acc.Phone, acc.Proxy, nil, db.Conn, acc.ID, dispatcher)
-	if err != nil {
-		return err
-	}
-
-	ctx := context.Background()
-	return client.Run(ctx, func(ctx context.Context) error {
-		api = tg.NewClient(client)
-		for _, cd := range dups {
-			donorUser, err := base.Modf_ExtractUsername(cd.URLChannelDonor)
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] некорректная ссылка %s: %v", cd.URLChannelDonor, err)
-				continue
-			}
-			donorResolved, err := api.ContactsResolveUsername(ctx, &tg.ContactsResolveUsernameRequest{Username: donorUser})
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] не удалось получить канал %s: %v", cd.URLChannelDonor, err)
-				continue
-			}
-			donorCh, err := base.Modf_FindChannel(donorResolved.GetChats())
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] канал %s не найден: %v", cd.URLChannelDonor, err)
-				continue
-			}
-			if err := base.Modf_JoinChannel(ctx, api, donorCh, db, acc.ID); err != nil && !strings.Contains(err.Error(), "USER_ALREADY_PARTICIPANT") {
-				log.Printf("[CHANNEL DUPLICATE] подписка на %s: %v", cd.URLChannelDonor, err)
-				continue
-			}
-			settings := tg.InputPeerNotifySettings{}
-			settings.SetMuteUntil(0)
-			_, err = api.AccountUpdateNotifySettings(ctx, &tg.AccountUpdateNotifySettingsRequest{
-				Peer:     &tg.InputNotifyPeer{Peer: &tg.InputPeerChannel{ChannelID: donorCh.ID, AccessHash: donorCh.AccessHash}},
-				Settings: settings,
-			})
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] уведомления %s: %v", cd.URLChannelDonor, err)
-			}
-			if cd.ChannelDonorTGID == nil {
-				_ = db.SetChannelDonorTGID(cd.ID, fmt.Sprintf("%d", donorCh.ID))
-			}
-
-			targetUser, err := base.Modf_ExtractUsername(cd.OrderURL)
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] некорректная ссылка %s: %v", cd.OrderURL, err)
-				continue
-			}
-			targetResolved, err := api.ContactsResolveUsername(ctx, &tg.ContactsResolveUsernameRequest{Username: targetUser})
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] не удалось получить канал %s: %v", cd.OrderURL, err)
-				continue
-			}
-			targetCh, err := base.Modf_FindChannel(targetResolved.GetChats())
-			if err != nil {
-				log.Printf("[CHANNEL DUPLICATE] канал %s не найден: %v", cd.OrderURL, err)
-				continue
-			}
-			if err := base.Modf_JoinChannel(ctx, api, targetCh, db, acc.ID); err != nil && !strings.Contains(err.Error(), "USER_ALREADY_PARTICIPANT") {
-				log.Printf("[CHANNEL DUPLICATE] подписка на %s: %v", cd.OrderURL, err)
-			}
-			if cd.OrderChannelTGID == nil {
-				_ = db.SetOrderChannelTGID(cd.OrderID, fmt.Sprintf("%d", targetCh.ID))
-			}
-			chMap[donorCh.ID] = channelInfo{id: cd.ID, donor: donorCh, target: targetCh}
+	for _, cd := range dups {
+		donorUser, err := base.Modf_ExtractUsername(cd.URLChannelDonor)
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] некорректная ссылка %s: %v", cd.URLChannelDonor, err)
+			continue
 		}
-		<-ctx.Done()
-		return nil
-	})
+		donorResolved, err := api.ContactsResolveUsername(ctx, &tg.ContactsResolveUsernameRequest{Username: donorUser})
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] не удалось получить канал %s: %v", cd.URLChannelDonor, err)
+			continue
+		}
+		donorCh, err := base.Modf_FindChannel(donorResolved.GetChats())
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] канал %s не найден: %v", cd.URLChannelDonor, err)
+			continue
+		}
+		if err := base.Modf_JoinChannel(ctx, api, donorCh, db, accountID); err != nil && !strings.Contains(err.Error(), "USER_ALREADY_PARTICIPANT") {
+			log.Printf("[CHANNEL DUPLICATE] подписка на %s: %v", cd.URLChannelDonor, err)
+			continue
+		}
+		settings := tg.InputPeerNotifySettings{}
+		settings.SetMuteUntil(0)
+		_, err = api.AccountUpdateNotifySettings(ctx, &tg.AccountUpdateNotifySettingsRequest{
+			Peer:     &tg.InputNotifyPeer{Peer: &tg.InputPeerChannel{ChannelID: donorCh.ID, AccessHash: donorCh.AccessHash}},
+			Settings: settings,
+		})
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] уведомления %s: %v", cd.URLChannelDonor, err)
+		}
+		if cd.ChannelDonorTGID == nil {
+			_ = db.SetChannelDonorTGID(cd.ID, fmt.Sprintf("%d", donorCh.ID))
+		}
+
+		targetUser, err := base.Modf_ExtractUsername(cd.OrderURL)
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] некорректная ссылка %s: %v", cd.OrderURL, err)
+			continue
+		}
+		targetResolved, err := api.ContactsResolveUsername(ctx, &tg.ContactsResolveUsernameRequest{Username: targetUser})
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] не удалось получить канал %s: %v", cd.OrderURL, err)
+			continue
+		}
+		targetCh, err := base.Modf_FindChannel(targetResolved.GetChats())
+		if err != nil {
+			log.Printf("[CHANNEL DUPLICATE] канал %s не найден: %v", cd.OrderURL, err)
+			continue
+		}
+		if err := base.Modf_JoinChannel(ctx, api, targetCh, db, accountID); err != nil && !strings.Contains(err.Error(), "USER_ALREADY_PARTICIPANT") {
+			log.Printf("[CHANNEL DUPLICATE] подписка на %s: %v", cd.OrderURL, err)
+		}
+		if cd.OrderChannelTGID == nil {
+			_ = db.SetOrderChannelTGID(cd.OrderID, fmt.Sprintf("%d", targetCh.ID))
+		}
+		chMap[donorCh.ID] = channelInfo{id: cd.ID, donor: donorCh, target: targetCh}
+	}
 }
