+++ pkg/telegram/module/monitoring/monitoring.go
@@ -12,7 +12,6 @@ import (
 	"atg_go/models"
 	"atg_go/pkg/storage"
 	base "atg_go/pkg/telegram/module"
-	accountmutex "atg_go/pkg/telegram/module/account_mutex"
 
 	"github.com/gotd/td/tg"
 )
@@ -41,39 +40,15 @@ func randomByPercent(base int, min, max float64) int {
 	return floor + 1
 }
 
-// Start запускает отслеживание новых постов на каналах заказов.
-// Используется первым доступным мониторинг-аккаунтом.
-func Start(db *storage.DB) {
-	go func() {
-		if err := run(db); err != nil {
-			log.Printf("[MONITORING] остановлено: %v", err)
-		}
-	}()
-}
-
-// run выполняет инициализацию клиента Telegram и обрабатывает обновления.
-func run(db *storage.DB) error {
-	rand.Seed(time.Now().UnixNano())
-	accounts, err := db.GetMonitoringAccounts()
-	if err != nil {
-		return err
-	}
-	if len(accounts) == 0 {
-		return fmt.Errorf("нет аккаунтов для мониторинга")
-	}
-	acc := accounts[0]
-
-	if err := accountmutex.LockAccount(acc.ID); err != nil {
-		return err
-	}
-	defer accountmutex.UnlockAccount(acc.ID)
-
+// Connect присоединяет модуль мониторинга к существующему клиенту Telegram.
+// Предполагается, что клиент и диспетчер уже инициализированы и работают.
+func Connect(ctx context.Context, api *tg.Client, dispatcher *tg.UpdateDispatcher, db *storage.DB, accountID int) {
 	orders, err := db.GetOrdersForMonitoring()
 	if err != nil {
-		return err
+		log.Printf("[MONITORING] получение заказов: %v", err)
+		return
 	}
 
-	dispatcher := tg.NewUpdateDispatcher()
 	orderMap := make(map[int64]orderInfo)
 
 	dispatcher.OnNewChannelMessage(func(ctx context.Context, e tg.Entities, upd *tg.UpdateNewChannelMessage) error {
@@ -167,53 +142,38 @@ func run(db *storage.DB) error {
 		return nil
 	})
 
-	client, err := base.Modf_AccountInitialization(acc.ApiID, acc.ApiHash, acc.Phone, acc.Proxy, nil, db.Conn, acc.ID, dispatcher)
-	if err != nil {
-		return err
-	}
-
-	ctx := context.Background()
-
-	return client.Run(ctx, func(ctx context.Context) error {
-		api := tg.NewClient(client)
-
-		// Подписываемся на каналы и включаем уведомления
-		for _, o := range orders {
-			username, err := base.Modf_ExtractUsername(o.URLDefault)
-			if err != nil {
-				log.Printf("[MONITORING] некорректная ссылка %s: %v", o.URLDefault, err)
-				continue
-			}
-			resolved, err := api.ContactsResolveUsername(ctx, &tg.ContactsResolveUsernameRequest{Username: username})
-			if err != nil {
-				log.Printf("[MONITORING] не удалось получить канал %s: %v", o.URLDefault, err)
-				continue
-			}
-			ch, err := base.Modf_FindChannel(resolved.GetChats())
-			if err != nil {
-				log.Printf("[MONITORING] канал %s не найден: %v", o.URLDefault, err)
-				continue
-			}
-			if err := base.Modf_JoinChannel(ctx, api, ch, db, acc.ID); err != nil && !strings.Contains(err.Error(), "USER_ALREADY_PARTICIPANT") {
-				log.Printf("[MONITORING] подписка на %s: %v", o.URLDefault, err)
-			}
-			settings := tg.InputPeerNotifySettings{}
-			settings.SetMuteUntil(0)
-			_, err = api.AccountUpdateNotifySettings(ctx, &tg.AccountUpdateNotifySettingsRequest{
-				Peer:     &tg.InputNotifyPeer{Peer: &tg.InputPeerChannel{ChannelID: ch.ID, AccessHash: ch.AccessHash}},
-				Settings: settings,
-			})
-			if err != nil {
-				log.Printf("[MONITORING] уведомления %s: %v", o.URLDefault, err)
-			}
-			if o.ChannelTGID == nil {
-				_ = db.SetOrderChannelTGID(o.ID, fmt.Sprintf("%d", ch.ID))
-			}
-			orderMap[ch.ID] = orderInfo{id: o.ID, url: o.URLDefault}
+	// Подписываемся на каналы заказов и включаем уведомления
+	for _, o := range orders {
+		username, err := base.Modf_ExtractUsername(o.URLDefault)
+		if err != nil {
+			log.Printf("[MONITORING] некорректная ссылка %s: %v", o.URLDefault, err)
+			continue
 		}
-
-		// держим соединение активным, пока контекст не будет отменён
-		<-ctx.Done()
-		return nil
-	})
+		resolved, err := api.ContactsResolveUsername(ctx, &tg.ContactsResolveUsernameRequest{Username: username})
+		if err != nil {
+			log.Printf("[MONITORING] не удалось получить канал %s: %v", o.URLDefault, err)
+			continue
+		}
+		ch, err := base.Modf_FindChannel(resolved.GetChats())
+		if err != nil {
+			log.Printf("[MONITORING] канал %s не найден: %v", o.URLDefault, err)
+			continue
+		}
+		if err := base.Modf_JoinChannel(ctx, api, ch, db, accountID); err != nil && !strings.Contains(err.Error(), "USER_ALREADY_PARTICIPANT") {
+			log.Printf("[MONITORING] подписка на %s: %v", o.URLDefault, err)
+		}
+		settings := tg.InputPeerNotifySettings{}
+		settings.SetMuteUntil(0)
+		_, err = api.AccountUpdateNotifySettings(ctx, &tg.AccountUpdateNotifySettingsRequest{
+			Peer:     &tg.InputNotifyPeer{Peer: &tg.InputPeerChannel{ChannelID: ch.ID, AccessHash: ch.AccessHash}},
+			Settings: settings,
+		})
+		if err != nil {
+			log.Printf("[MONITORING] уведомления %s: %v", o.URLDefault, err)
+		}
+		if o.ChannelTGID == nil {
+			_ = db.SetOrderChannelTGID(o.ID, fmt.Sprintf("%d", ch.ID))
+		}
+		orderMap[ch.ID] = orderInfo{id: o.ID, url: o.URLDefault}
+	}
 }
